edit c++.sh to return error

+++ c++.sh
@@ -1,11 +1,17 @@
-#!/bin/zsh
+#!/bin/sh
 
-# compile .cpp file as temp.out by clang++
-clang++ $1 -o temp.out
+# return error when it gets too many args...
+if [ $# -ge 2 ]
+  then
+    echo "An error occured. Too many args."
+else
+  # compile .cpp file as temp.out by clang++
+  clang++ $1 -o temp.out
 
-# run compiled binary file
-./temp.out
+  # run compiled binary file
+  ./temp.out
 
-# delete compiled binary file
-rm temp.out
+  # delete compiled binary file
+  rm temp.out
+fi
